requirements.txt
transformers
scikit-learn
datasets
torch
accelerate
seqeval
pyyaml